move code in a separate controller
