+++ apps/transport/lib/transport_web/controllers/validation_controller.ex
@@ -42,29 +42,6 @@ defmodule TransportWeb.ValidationController do
     end
   end
 
-  def convert(%Plug.Conn{} = conn, %{"upload" => upload_params}) do
-    file_path = upload_params["file"].path
-
-    conn =
-      case call_geojson_converter(file_path) do
-        {:ok, res} ->
-          conn
-          |> Plug.Conn.resp(200, res)
-        {:error, err_msg} -> conn |> Plug.Conn.resp(400, err_msg)
-      end
-
-    conn
-    |> Plug.Conn.send_resp()
-  end
-
-  def call_geojson_converter(file_path) do
-    case Rambo.run("gtfs-geojson", ["--input", file_path]) do
-      {:ok, %Rambo{out: res}} -> {:ok, res}
-      {:error, %Rambo{err: err_msg}} -> {:error, err_msg}
-      {:error, _} = r -> r
-    end
-  end
-
   def show(%Plug.Conn{} = conn, %{} = params) do
     config = make_pagination_config(params)
     validation = Repo.get(Validation, params["id"])
