+++ apps/transport/lib/transport_web/templates/geojson_conversion/index.html.eex
@@ -1,8 +1,8 @@
 <section class="single-page section section-grey">
   <div class="container">
     <article class="panel">
-    <h1><%= TransportWeb.Gettext.dgettext("validations", "Check a GTFS file's quality") %></h1>
-      <%= form_for @conn, validation_path(@conn, :validate), [as: "upload", multipart: :true], fn f -> %>
+    <h1>Convertir un GTFS en GeoJson</h1>
+      <%= form_for @conn, geojson_conversion_path(@conn, :convert), [as: "upload", multipart: :true], fn f -> %>
         <%= file_input f, :file, accept: ".zip" %>
         <%= submit dgettext("validations", "Upload") %>
       <% end %>
