+++ apps/transport/lib/transport_web/views/geojson_conversion_view.ex
@@ -1,0 +1,3 @@
+defmodule TransportWeb.GeojsonConversionView do
+  use TransportWeb, :view
+end
